# ms-gateway/docs/swagger.yaml
definitions:
  helper.Message:
    properties:
      message:
        type: string
    type: object
  helper.MessageDetails:
    properties:
      detail:
        type: string
      message:
        type: string
    type: object
...
  model.CreateHotelRequest:
    properties:
      description:
        type: string
      location:
        type: string
      name:
        type: string
    type: object
...
  model.CreateRoomRequest:
    properties:
      capacity:
        type: integer
      hotel_id:
        type: integer
      price:
        type: integer
    type: object
  model.Hotel:
    properties:
      create_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      update_at:
        type: string
    type: object
...
  model.Room:
    properties:
      capacity:
        type: integer
      create_at:
        type: string
      hotel_id:
        type: integer
      id:
        type: integer
      price:
        type: integer
      update_at:
        type: string
    type: object
...
  model.UpdateRoomRequest:
    properties:
      capacity:
        type: integer
      hotel_id:
        type: integer
      id:
        type: integer
      price:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: arviansyahnur3@gmail.com
    name: Arviansyah Nur
    url: http://www.swagger.io/support
  description: An e-commerce API for booking hotel and with many features, made using
    microservices and gRPC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: vihotel
  version: "1.0"
paths:
...
  /api/v1/hotel:
    get:
      consumes:
      - application/json
      description: Get all Hotel
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Hotel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Public
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create Hotel
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CreateHotelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Hotel
  /api/v1/hotel/:id:
    delete:
      consumes:
      - application/json
      description: Delete Hotel
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Hotel
    get:
      consumes:
      - application/json
      description: Get info Hotel
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Hotel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Public
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update Hotel
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CreateHotelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Hotel
  /api/v1/hotel/room:
    get:
      consumes:
      - application/json
      description: Get all room
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Public
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create room
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Hotel
    put:
      consumes:
      - application/json
      description: Update room
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Hotel
  /api/v1/hotel/room/:id:
    delete:
      consumes:
      - application/json
      description: Delete room
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Hotel
    get:
      consumes:
      - application/json
      description: Get info room
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Public
      tags:
      - User
...
swagger: "2.0"